+++ lib/has_config.rb
@@ -1,10 +1,11 @@
 require 'has_config/version'
-require 'has_config/config_reader'
-require 'has_config/engine'
 require 'has_config/errors'
-require 'has_config/config_setting'
+require 'has_config/configuration'
 require 'has_config/value_parser'
-require 'has_config/active_record_model'
+require 'has_config/chain'
+require 'has_config/engine'
+require 'has_config/active_record/processor'
+require 'has_config/active_record/model_adapter'
 
 module HasConfig
 end