+++ lib/has_config/errors.rb
@@ -1,6 +1,7 @@
 module HasConfig
-  ConfigurationNotFound = Class.new(StandardError)
-  InvalidChainOption    = Class.new(StandardError)
-  InvalidType           = Class.new(StandardError)
-  UnknownConfig         = Class.new(StandardError)
+  ConfigurationFileNotFound = Class.new(StandardError)
+  InvalidChainOption        = Class.new(StandardError)
+  InvalidChain              = Class.new(StandardError)
+  InvalidType               = Class.new(StandardError)
+  UnknownConfig             = Class.new(StandardError)
 end