+++ lib/has_config/value_parser.rb
@@ -8,18 +8,7 @@ module HasConfig
       when :integer
         value.present? ? value.to_i : nil
       when :boolean
-        [true, 1].include?(value) || /\A(true|t|yes|y|1)\z/i === value
-      end
-    end
-
-    def self.invoke_chain?(value, chain_on)
-      case chain_on
-      when :blank
-        value.blank?
-      when :nil
-        value.nil?
-      when :false
-        value == false
+        /\A(true|t|yes|y|1)\z/i === value.to_s
       end
     end
   end