+++ test/config/has_config.rb
@@ -1,0 +1,5 @@
+has_config :string_setting, config: { type: :string }
+has_config :boolean_setting, config: { type: :boolean }
+has_config :integer_setting, config: { type: :integer }
+
+has_config :chained_integer, config: { type: :integer }