+++ test/config/settings.rb
@@ -1,9 +1,0 @@
-config :favorite_color, :string
-config :enable_email, :boolean
-config :rate_limit, :integer
-
-config :listed_rate_limit, :integer, validations: [{ inclusion: { in: [1, 2, 3] } }, { presence: true }]
-config :required_favorite_color, :string, validations: { presence: true }
-config :favorite_color_default, :string, default: 'red'
-
-config :chained_config, :integer