+++ test/models.rb
@@ -1,66 +1,86 @@
+class BasicModel < ActiveRecord::Base
+  serialize :configuration
+  include HasConfig::ActiveRecord::ModelAdapter
+
+  has_config :string_setting
+  has_config :boolean_setting
+  has_config :integer_setting
+
+  def bad_parent_method
+    nil
+  end
+end
+
+class ManualSettingModel < ActiveRecord::Base
+  self.table_name = 'basic_models'
+  serialize :configuration
+  include HasConfig::ActiveRecord::ModelAdapter
+
+  has_config :string_setting, config: { default: 'custom' }
+  has_config :manual_setting, config: { type: :string, default: 'manual' }
+end
+
+class WithValidationModel < ActiveRecord::Base
+  self.table_name = 'basic_models'
+  serialize :configuration
+  include HasConfig::ActiveRecord::ModelAdapter
+
+  has_config :string_setting, config: { type: :string, validations: { presence: true } }
+end
+
 class CustomColumnModel < ActiveRecord::Base
   serialize :prefs
-  include HasConfig::ActiveRecordModel
-  self.config_column = :prefs
+  include HasConfig::ActiveRecord::ModelAdapter
+  self.has_config_configuration_column = :prefs
 
-  has_config :favorite_color
-  has_config :enable_email
-  has_config :rate_limit
+  has_config :string_setting
+  has_config :boolean_setting
+  has_config :integer_setting
 end
 
-class HashModel < ActiveRecord::Base
+class JsonColumnModel < ActiveRecord::Base
+  include HasConfig::ActiveRecord::ModelAdapter
+
+  has_config :string_setting
+  has_config :boolean_setting
+  has_config :integer_setting
+end
+
+class JsonbColumnModel < ActiveRecord::Base
+  include HasConfig::ActiveRecord::ModelAdapter
+
+  has_config :string_setting
+  has_config :boolean_setting
+  has_config :integer_setting
+end
+
+class Client < ActiveRecord::Base
+  has_many :groups
+  include HasConfig::ActiveRecord::ModelAdapter
+  has_config :chained_integer
+end
+
+class Group < ActiveRecord::Base
+  belongs_to :client
+  has_many :users
+  include HasConfig::ActiveRecord::ModelAdapter
+  has_config :chained_integer, parent: :client
+end
+
+class User < ActiveRecord::Base
+  belongs_to :group
+  include HasConfig::ActiveRecord::ModelAdapter
+  has_config :chained_integer, parent: :group
+end
+
+class ParentModel < ActiveRecord::Base
   serialize :configuration
-  include HasConfig::ActiveRecordModel
+  include HasConfig::ActiveRecord::ModelAdapter
 
-  has_config :favorite_color
-  has_config :enable_email
-  has_config :rate_limit
+  has_config :string_setting
+  has_config :boolean_setting
+  has_config :integer_setting
 end
 
-class JsonModel < ActiveRecord::Base
-  include HasConfig::ActiveRecordModel
-
-  has_config :favorite_color
-  has_config :enable_email
-  has_config :rate_limit
-end
-
-class WithDefault < ActiveRecord::Base
-  serialize :configuration
-  include HasConfig::ActiveRecordModel
-
-  has_config :favorite_color
-  has_config :enable_email
-  has_config :rate_limit
-
-  has_config :favorite_color_default
-end
-
-class WithValidation < ActiveRecord::Base
-  serialize :configuration
-  include HasConfig::ActiveRecordModel
-
-  has_config :favorite_color
-  has_config :enable_email
-  has_config :rate_limit
-
-  has_config :listed_rate_limit
-  has_config :required_favorite_color
-end
-
-class ChainOne < ActiveRecord::Base
-  include HasConfig::ActiveRecordModel
-  has_config :chained_config, parent: :chain_two
-  belongs_to :chain_two
-end
-
-class ChainTwo < ActiveRecord::Base
-  include HasConfig::ActiveRecordModel
-  has_config :chained_config, parent: :chain_three
-  belongs_to :chain_three
-end
-
-class ChainThree < ActiveRecord::Base
-  include HasConfig::ActiveRecordModel
-  has_config :chained_config
+class ChildModel < ParentModel
 end