+++ test/schema.rb
@@ -1,37 +1,43 @@
 ActiveRecord::Schema.define do
   self.verbose = false
 
+  create_table :basic_models, force: true do |t|
+    t.text :configuration
+  end
+
+  create_table :parent_models, force: true do |t|
+    t.string :type
+    t.text :configuration
+  end
+
   create_table :custom_column_models, force: true do |t|
     t.text :prefs
   end
 
-  create_table :hash_models, force: true do |t|
-    t.text :configuration
-  end
-
-  create_table :json_models, force: true do |t|
+  create_table :json_column_models, force: true do |t|
     t.json :configuration
   end
 
-  create_table :with_defaults, force: true do |t|
-    t.text :configuration
+  create_table :jsonb_column_models, force: true do |t|
+    # JSONB support was added in Rails 4.2
+    if ActiveRecord.version >= Gem::Version.new('4.2.0')
+      t.jsonb :configuration
+    else
+      t.json :configuration
+    end
   end
 
-  create_table :with_validations, force: true do |t|
-    t.text :configuration
+  create_table :clients, force: true do |t|
+    t.json :configuration
   end
 
-  create_table :chain_ones, force: true do |t|
-    t.integer :chain_two_id
-    t.json :configuration, default: {}
+  create_table :groups, force: true do |t|
+    t.integer :client_id
+    t.json :configuration
   end
 
-  create_table :chain_twos, force: true do |t|
-    t.integer :chain_three_id
-    t.json :configuration, default: {}
-  end
-
-  create_table :chain_threes, force: true do |t|
-    t.json :configuration, default: {}
+  create_table :users, force: true do |t|
+    t.integer :group_id
+    t.json :configuration
   end
 end