+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0 (Unreleased)
+
+* Full rewrite
+* Allow predefined settings in a configuration file
+* Allow chaining from models of the same setting
+* Removed grouping functionality
+
 ## 0.1.0 (6/20/2015)
 
 * Initial release